[step03][refactor] PR feedback applied

+++ src/main/java/nextstep/courses/domain/ImageRepository.java
@@ -1,7 +1,7 @@
 package nextstep.courses.domain;
 
 public interface ImageRepository {
-    int save(Image image);
+    Long save(Image image);
 
     Image findById(Long id);
 }
