+++ src/main/java/nextstep/courses/domain/SessionEnrollRepository.java
@@ -1,7 +1,7 @@
 package nextstep.courses.domain;
 
 public interface SessionEnrollRepository {
-    int save(SessionEnroll sessionEnroll);
+    Long save(SessionEnroll sessionEnroll);
 
     SessionEnroll findById(Long id);
 }
